requirements.txt
# Essential packages - Updated for Python 3.13 compatibility
# Django 5.1.x supports Python 3.13
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
python-decouple==3.8
mysqlclient==2.2.6
pytest==8.3.4
pytest-django==4.9.0
google-ads==28.0.0
google-ads-admanager==0.2.5
googleads==41.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0
PyYAML==6.0.2
requests==2.32.3
django-crontab==0.7.1
